restructure dto
